Sets up Spring Framework

+++ src/main/java/pint/User.java
@@ -2,7 +2,7 @@ package pint;
 
 public class User {
 
-    private final int id;
+    private final long id;
     private final String username;
     private final String email;
